Move tests to own directory
